+++ core.impl/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/impl/repair/state/RepairStateImpl.java
@@ -87,38 +87,53 @@ public class RepairStateImpl implements RepairState
     @Override
     public final void update()
     {
-        RepairStateSnapshot oldRepairStateSnapshot = myRepairStateSnapshot.get();
-        long now = System.currentTimeMillis();
-        if (oldRepairStateSnapshot == null
-                || isRepairNeeded(oldRepairStateSnapshot.lastCompletedAt(),
-                oldRepairStateSnapshot.getEstimatedRepairTime(),
-                now))
+        try
         {
-            RepairStateSnapshot newRepairStateSnapshot = generateNewRepairState(myNode, oldRepairStateSnapshot, now);
-            if (myRepairStateSnapshot.compareAndSet(oldRepairStateSnapshot, newRepairStateSnapshot))
+            RepairStateSnapshot oldRepairStateSnapshot = myRepairStateSnapshot.get();
+            long now = System.currentTimeMillis();
+            if (oldRepairStateSnapshot == null
+                    || isRepairNeeded(oldRepairStateSnapshot.lastCompletedAt(),
+                    oldRepairStateSnapshot.getEstimatedRepairTime(),
+                    now))
             {
-                myTableRepairMetrics.lastRepairedAt(myTableReference, newRepairStateSnapshot.lastCompletedAt());
+                RepairStateSnapshot newRepairStateSnapshot = generateNewRepairState(myNode,
+                        oldRepairStateSnapshot, now);
+                if (myRepairStateSnapshot.compareAndSet(oldRepairStateSnapshot, newRepairStateSnapshot))
+                {
+                    myTableRepairMetrics.lastRepairedAt(myTableReference,
+                            newRepairStateSnapshot.lastCompletedAt());
 
-                int nonRepairedRanges
-                        = (int) newRepairStateSnapshot.getVnodeRepairStates().getVnodeRepairStates().stream()
-                        .filter(v -> vnodeIsRepairable(v, newRepairStateSnapshot, System.currentTimeMillis()))
-                        .count();
+                    int nonRepairedRanges
+                            = (int) newRepairStateSnapshot.getVnodeRepairStates().getVnodeRepairStates()
+                            .stream()
+                            .filter(v -> vnodeIsRepairable(v, newRepairStateSnapshot,
+                                    System.currentTimeMillis()))
+                            .count();
 
-                int repairedRanges
-                        = newRepairStateSnapshot.getVnodeRepairStates().getVnodeRepairStates().size()
-                        - nonRepairedRanges;
-                myTableRepairMetrics.repairState(myTableReference, repairedRanges, nonRepairedRanges);
-                myTableRepairMetrics.remainingRepairTime(myTableReference,
-                        newRepairStateSnapshot.getRemainingRepairTime(System.currentTimeMillis(),
-                                myRepairConfiguration.getRepairIntervalInMs()));
-                LOG.trace("Table {} switched to repair state {}", myTableReference, newRepairStateSnapshot);
+                    int repairedRanges
+                            = newRepairStateSnapshot.getVnodeRepairStates().getVnodeRepairStates().size()
+                            - nonRepairedRanges;
+                    myTableRepairMetrics.repairState(myTableReference, repairedRanges, nonRepairedRanges);
+                    myTableRepairMetrics.remainingRepairTime(myTableReference,
+                            newRepairStateSnapshot.getRemainingRepairTime(System.currentTimeMillis(),
+                                    myRepairConfiguration.getRepairIntervalInMs()));
+                    LOG.trace("Table {} switched to repair state {}", myTableReference,
+                            newRepairStateSnapshot);
+                }
+            }
+            else
+            {
+                LOG.trace("Table {} keeping repair state {}", myTableReference, oldRepairStateSnapshot);
             }
         }
-        else
+        finally
         {
-            LOG.trace("Table {} keeping repair state {}", myTableReference, oldRepairStateSnapshot);
+            RepairStateSnapshot snapshot = myRepairStateSnapshot.get();
+            if (snapshot != null)
+            {
+                myPostUpdateHook.postUpdate(snapshot);
+            }
         }
-        myPostUpdateHook.postUpdate(myRepairStateSnapshot.get());
     }
 
     /**